Better handling of Wayland compatibility mode.
- Factor wayland detection off into a module and use it to conditionally call the new XInitThreads.
- Remove the incorrect detection of GDK_BACKEND=x11 warning.
