+++ ulauncher/main.py
@@ -20,9 +20,9 @@ import dbus
 import dbus.service
 from dbus.mainloop.glib import DBusGMainLoop
 
-from ulauncher.config import (get_version, get_options, is_wayland, is_wayland_compatibility_on,
-                              gdk_backend, CACHE_DIR, CONFIG_DIR, DATA_DIR)
+from ulauncher.config import get_version, get_options, CACHE_DIR, CONFIG_DIR, DATA_DIR
 from ulauncher.utils.decorator.run_async import run_async
+from ulauncher.utils.wayland import is_wayland, is_wayland_compatibility_on
 from ulauncher.ui.windows.UlauncherWindow import UlauncherWindow
 from ulauncher.ui.AppIndicator import AppIndicator
 from ulauncher.utils.Settings import Settings
