+++ ulauncher/ui/windows/PreferencesUlauncherDialog.py
@@ -33,12 +33,13 @@ from ulauncher.utils.Theme import themes, Theme, load_available_themes
 from ulauncher.utils.decorator.glib_idle_add import glib_idle_add
 from ulauncher.utils.mypy_extensions import TypedDict
 from ulauncher.utils.decorator.run_async import run_async
+from ulauncher.utils.wayland import is_wayland
 from ulauncher.utils.Settings import Settings
 from ulauncher.utils.Router import Router, get_url_params
 from ulauncher.utils.AutostartPreference import AutostartPreference
 from ulauncher.ui.AppIndicator import AppIndicator
 from ulauncher.search.shortcuts.ShortcutsDb import ShortcutsDb
-from ulauncher.config import get_data_file, get_options, get_version, is_wayland, EXTENSIONS_DIR
+from ulauncher.config import get_data_file, get_options, get_version, EXTENSIONS_DIR
 
 
 logger = logging.getLogger(__name__)
